feat: Update last_message and last_timestamp in chat3controller
Update the values of last_message and last_timestamp in the
chat3controller's index function. Additionally, include the Title
property from the data object. This change ensures that the latest
message and timestamp are accurately recorded and can be used in the
application logic. It improves the accuracy and relevancy of the
displayed chat information.