Convert DALL-E PNG images to JPG in public/img folder
This updates the setup.js file to add functionality for converting PNG
images to JPG format in the public/img folder. The conversion process is
triggered when a PNG file exists without a corresponding JPG file in the
folder. The converted JPG images have a quality setting of 70. This
change ensures that all images in the public/img folder are in the
desired JPG format, improving compatibility and reducing file size. No
known issues are introduced with this change.